+++ UnitTests/BraintreeVenmoTests/BTVenmoRequest_Tests.swift
@@ -18,4 +18,14 @@ class BTVenmoRequest_Tests: XCTestCase {
         XCTAssertEqual(request.collectCustomerShippingAddress, false)
         XCTAssertEqual(request.collectCustomerBillingAddress, false)
     }
+
+    func testIsFinalAmount_whenIsFinalAmountNotSet_defaultsToFalse() {
+        let request = BTVenmoRequest(paymentMethodUsage: .singleUse)
+        XCTAssertEqual(request.isFinalAmount, false)
+    }
+
+    func testFallbackToWeb_whenNotPassed_defaultsValueAsFalse() {
+        let request = BTVenmoRequest(paymentMethodUsage: .singleUse)
+        XCTAssertEqual(request.fallbackToWeb, false)
+    }
 }